refactor: 버킷 달성 현황 컴포넌트와 달성 이미지 컴포넌트를 분리하여 관리
