+++ src/components/Home/BucketAchievement/AchievementImage/AchievementImage.style.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  position: absolute;
+  left: 10rem;
+  top: 10rem;
+`;
