+++ src/components/Home/BucketAchievement/BucketAchievement.style.ts
@@ -40,8 +40,3 @@ export const BucketlistLabel = styled.span`
   top: 2rem;
   left: 2.5rem;
 `;
-export const IconWrapper = styled.div`
-  position: absolute;
-  left: 10rem;
-  top: 10rem;
-`;
